chore: Supports meta-magical example-based testing
This allows examples to be provided in doc comments, and for these examples to be tested in Mocha. Testable examples are those that are marked with `::` in the preceding paragraph of the documentation. Headings that end with `::` can provide a distinct name to the example that follows it:

    /*~
     * ## Using identity::
     *
     *     identity(1)  // ==> 1
     */
    const identity = (x) => x

Note that assertions are described with assertion comments (`// ==> expression`). Deep structural equality is checked when running the assertions.

To create the assertions in Mocha, in your Mocha file:

    const metamagical = require('metamagical-interface');
    const defineTests = require('metamagical-mocha-bridge')(metamagical, describe, it);
    const _ = require('../').your.module;

    describe('Your.Module', function() {
      defineTests(_)
    })