feat: Cleaner function-based ADT API
This moves to boris' suggested function-based API for ADTs. Normalisation and type checking with recursive types are trivially supported by just doing the checks in the initialiser function itself (which can be improved with contracts, eventually).

This also adds a `.hasInstance()` method to each ADT, that checks for instances structurally. This would use JS's native `Symbol.hasInstance`, which is used by `a instanceof b` in ES2015, but we can't reliably have the same behaviour in all engines, which would make this more confusing.