(build docs) Support Guide-ish documents in the documentation
This adds a new special annotation, `@guide: <TITLE>` that supports
adding guides in the annotated documentation tree. Guides mutate the
documentation root (in this case, `folktale`'s library entrypoint), but
do so only when we're shipping the documentation, so the changes are
not visible for people getting only the library package.

References are still not handled in an ideal way, but links between guides
can just use regular HTML links for now.