(build docs) Support nesting guides in other documents
A `parent:` metadata may now be provided for guides, and is used
specifically to nest guides in the static-docs tool.

The value of `parent` is a JS expression that resolves to a
JavaScript object. The guide will be added to that object,
under the guide's name/title key. If not provided, `folktale`'s
entry point is assumed.