Fixes deferred state message.
When we try to settle a deferred that's already settled we get a message saying that it's not possible to move to the new state, because the current state is not "pending"... except it described the new state where it should describe the old one. Ooops!

This addresses part of the last comment in #153